Refactor UpdateName component and improve i18n support
